+++ src/lightning/fabric/strategies/xla_fsdp.py
@@ -599,7 +599,7 @@ class XLAFSDPStrategy(ParallelStrategy, _Sharded):
 
     @classmethod
     def register_strategies(cls, strategy_registry: _StrategyRegistry) -> None:
-        strategy_registry.register("xla_fsdp", cls, description=cls.__class__.__name__)
+        strategy_registry.register("xla_fsdp", cls, description=cls.__name__)
 
     def _parse_fsdp_kwargs(self) -> Dict:
         # this needs to be delayed because `self.precision` isn't available at init